requirements.txt
pandas
matplotlib
redis
grpcio
grpcio-tools
